fix(RouteEngine): Node and Linkedlist naming and distance type fixed

+++ RouteEngine/Core/Node.cpp
@@ -32,7 +32,7 @@ bool Node::GetIsObstacle() {
     return isObstacle;
 }
 
-float Node::GetDistance() {
+int Node::GetDistance() {
     return distance;
 }
